requirements.txt
ipinfo_django
django-ratelimit
redis
django-ipware
scikit-learn
django-ip-geolocation
celery